Require controller module in lib/prototypical.rb
By requiring the controller module in lib/prototypical.rb it gets
required when the gem itself is required, which for Rails apps will
typically happen when the app boots.

As a result it is no longer necessary to explicitly require the module
into the controllers that use it. This seems more in line with expected
behaviour for a Rails plugin.